requirements.txt
spotipy
openai
python-dotenv
streamlit
